Remove CCNS from code base. Issue #64
